# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Koishi API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://koishi-api.nechapu.to",
      "description": "The official Koishi API server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Koishi API running on localhost"
    }
  ],
  "components": {
    "schemas": {
      "ObjectId": {
        "type": "string",
        "example": "625401f6ffd69c32c336e0ee"
      },
...
      "Collection": {
        "type": "object",
        "description": "A base structure of a data collection",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "title": {
            "type": "string",
            "example": "Server monitoring"
          },
          "owner": {
            "description": "ObjectID referencing the collection's owner",
            "$ref": "#/components/schemas/ObjectId"
          },
          "isPublic": {
            "type": "boolean"
          },
          "sharedTo": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "userEmail": {
                  "type": "string",
                  "example": "guy@gmail.com"
                },
                "role": {
                  "type": "string",
                  "enum": ["view", "edit"]
                }
              }
            }
          },
          "model": {
            "description": "ObjectID referencing the collection's model",
            "$ref": "#/components/schemas/ObjectId"
          },
          "data": {
            "description": "ObjectID referencing the collection's data set",
            "$ref": "#/components/schemas/ObjectId"
          },
          "settings": {
            "type": "array",
            "description": "An array of ObjectIDs referencing the collection's settings presets",
            "items": {
              "type": "object",
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "actions": {
            "description": "ObjectID referencing the collection's actions module",
            "$ref": "#/components/schemas/ObjectId"
          }
        }
      },
...
      "ActionToken": {
        "description": "Used to confirm actions that may have fatal impact",
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "category": {
            "type": "string",
            "description": "For now only sharing and deleting collections",
            "enum": ["delete", "share"]
          },
          "purpose": {
            "type": "string",
            "description": "A string explaining the purpose of the action token",
            "example": "Delete collection \"Test collection\""
          },
          "userId": {
            "$ref": "#/components/schemas/ObjectId",
            "description": "References a user who issued this action token"
          },
          "targetId": {
            "$ref": "#/components/schemas/ObjectId",
            "description": "References whatever is the action token's target (depends on the use case)"
          }
        }
      }
    },
    "securitySchemes": {
      "authToken": {
        "type": "apiKey",
        "name": "x-auth-token",
        "in": "header"
      }
    }
  },
  "paths": {
...
    "/collections": {
      "get": {
        "description": "Get all collections",
        "tags": ["Collections"],
        "security": [
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of all collections you own or you are subscribed to.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Collection"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a new collection",
        "tags": ["Collections"],
        "security": [
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Collection title",
                    "example": "Server monitoring"
                  },
                  "model": {
                    "type": "array",
                    "description": "A list of columns defining the collection data structure",
                    "items": {
                      "type": "object",
                      "properties": {
                        "columnName": {
                          "type": "string",
                          "example": "CPU Usage"
                        },
                        "dataType": {
                          "type": "string",
                          "enum": ["number", "text", "date", "time", "bool"]
                        },
                        "unit": {
                          "type": "string",
                          "example": "%",
                          "nullable": true
                        }
                      } 
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {

              }
            }
          },
          "400": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/collections/{id}": {
      "get": {
        "tags": ["Collections"],
        "description": "Get a collection specified by it's id (ObjectID)",
        "parameters": [
          {
            "name": "id",
            "description": "ObjectID referencing a collection",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "noPopulate",
            "description": "By default, the returned collection has all its fields populated - that may lead into the returned json being very large. Setting ?noPopulate to \"true\" will return the collection without it's fields populated.",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean",
              "example": "false"
            }
          }
        ],
        "security": [
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found."
          }
        }
      },
      "put": {
        "description": "Change a collection's title",
        "tags": ["Collections"],
        "security": [
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "description": "ObjectID referencing a collection",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "New name"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection's title changed (returns the updated collection)"
          },
          "404": {
            "description": "Collection not found"
          }
        }
      },
      "delete": {
        "description": "Delete a collection. At first, you only supply the target collection's id. Nothing gets deleted, you only recieve an action token, that you then use to confirm the action (deleting a collection) by supplying it via the \"actionToken\" parameter in query.",
        "tags": ["Collections"],
        "parameters": [
          {
            "name": "id",
            "description": "ObjectID referencing a collection",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "actionToken",
            "description": "ID of the action token that confirms deleting a collection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "security": [
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a token that is then used to confirm deleting the collection. (Or then confirms that the collection is deleted)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionToken"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found"
          }
        }
      }
    }
  }
}